+++ packages/gpu_voxels/src/gpu_voxels/octree/RobotNodes.h
@@ -212,7 +212,7 @@ public:
     assert(
         uint64_t(child) != 0
             && uint64_t(child) < (uint64_t(1) << uint64_t(8 * (sizeof(m_child_high) + sizeof(m_child_low)))));
-    m_child_high = (uint8_t) (uint64_t(child) >> (8 * sizeof(m_child_low)));
+    m_child_high = (uint16_t) (uint64_t(child) >> (8 * sizeof(m_child_low)));
     m_child_low = (uint32_t) (uint64_t(child) & 0xFFFFFFFF);
   }
 
@@ -229,7 +229,7 @@ public:
 
 private:
   uint32_t m_child_low;
-  uint8_t m_child_high;
+  uint16_t m_child_high;
   uint8_t m_status;
   uint16_t alignment;
   BitVector<BIT_VECTOR_LENGTH> m_voxel_meaning;
